Chore: Remastered CLI using ConcurrentLinkedQueue with ReentrantLock
