+++ CLI_Final/.idea/libraries/log4j_api_2_12_4.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="log4j-api-2.12.4">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/../../../../../../apache-log4j-2.12.4-bin/apache-log4j-2.12.4-bin/log4j-api-2.12.4.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
