+++ CLI_Final/Vendors.json
@@ -1,0 +1,10 @@
+[
+  {
+    "vendorId": "Hesh678",
+    "password": "ghjknm12"
+  },
+  {
+    "vendorId": "shal345",
+    "password": "wertyu12"
+  }
+]
