+++ CLI_Final/customer.json
@@ -1,0 +1,10 @@
+[
+  {
+    "customerId": "fren456",
+    "password": "dfghjk12"
+  },
+  {
+    "customerId": "gern234",
+    "password": "wsdfgh12"
+  }
+]
