--- RemasteredCLI/CLI/Tickets.json
@@ -1,232 +1,0 @@
-[
-  {
-    "ticketId": "gren234-1",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Booked",
-    "customerId": "robe234"
-  },
-  {
-    "ticketId": "jren234-1",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Booked",
-    "customerId": "relo234"
-  },
-  {
-    "ticketId": "kren234-1",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Booked",
-    "customerId": "bele234"
-  },
-  {
-    "ticketId": "gren234-1",
-    "eventName": "holo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Booked",
-    "customerId": "robe234"
-  },
-  {
-    "ticketId": "kren234-2",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Booked",
-    "customerId": "bele234"
-  },
-  {
-    "ticketId": "jren234-2",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-2",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Booked",
-    "customerId": "robe234"
-  },
-  {
-    "ticketId": "gren234-2",
-    "eventName": "holo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "kren234-3",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-3",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "jren234-3",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-3",
-    "eventName": "holo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "kren234-4",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "jren234-4",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-4",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-4",
-    "eventName": "holo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "kren234-5",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "jren234-5",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-5",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-5",
-    "eventName": "holo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "jren234-6",
-    "eventName": "yolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "jren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "kren234-6",
-    "eventName": "polo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "kren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  },
-  {
-    "ticketId": "gren234-6",
-    "eventName": "rolo",
-    "price": 75.3,
-    "timeDuration": "2 hours",
-    "Date": "2022-10-20",
-    "vendorId": "gren234",
-    "ticketStatus": "Available",
-    "customerId": "Not Set"
-  }
-]
