--- RemasteredCLI/CLI/Vendors.json
@@ -1,6 +1,0 @@
-[
-  {
-    "vendorId": "hesh234",
-    "password": "ehsjwk34"
-  }
-]
