--- RemasteredCLI/CLI/customer.json
@@ -1,10 +1,0 @@
-[
-  {
-    "customerId": "john202",
-    "password": "23gn45nn"
-  },
-  {
-    "customerId": "ehsn432",
-    "password": "fhgnfh23"
-  }
-]
