--- RemasteredCLI/CLI/logs/Customer_run.log
@@ -1,5 +1,0 @@
-2024-11-23 17:36:21 [Thread-4] INFO  CustomerRun - Ticket purchased by robe234 (Event: rolo)
-2024-11-23 17:36:21 [Thread-5] INFO  CustomerRun - Ticket purchased by bele234 (Event: polo)
-2024-11-23 17:36:21 [Thread-6] INFO  CustomerRun - Ticket purchased by relo234 (Event: yolo)
-2024-11-23 17:36:21 [Thread-7] INFO  CustomerRun - Ticket purchased by robe234 (Event: holo)
-2024-11-23 17:36:22 [Thread-4] INFO  CustomerRun - Ticket purchased by robe234 (Event: rolo)
