--- RemasteredCLI/CLI/logs/ticketPool_add_tickets.log
@@ -1,23 +1,0 @@
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-1 rolo
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-1 yolo
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-1 polo
-2024-11-23 17:36:20 [Thread-3] INFO  TicketPoolAdd - Ticket added to pool: gren234-1 holo
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-2 polo
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-2 yolo
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-2 rolo
-2024-11-23 17:36:20 [Thread-3] INFO  TicketPoolAdd - Ticket added to pool: gren234-2 holo
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-3 polo
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-3 rolo
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-3 yolo
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolAdd - Ticket added to pool: gren234-3 holo
-2024-11-23 17:36:21 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-4 polo
-2024-11-23 17:36:21 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-4 yolo
-2024-11-23 17:36:21 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-4 rolo
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolAdd - Ticket added to pool: gren234-4 holo
-2024-11-23 17:36:21 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-5 polo
-2024-11-23 17:36:21 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-5 yolo
-2024-11-23 17:36:21 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-5 rolo
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolAdd - Ticket added to pool: gren234-5 holo
-2024-11-23 17:36:22 [Thread-2] INFO  TicketPoolAdd - Ticket added to pool: jren234-6 yolo
-2024-11-23 17:36:22 [Thread-1] INFO  TicketPoolAdd - Ticket added to pool: kren234-6 polo
-2024-11-23 17:36:22 [Thread-0] INFO  TicketPoolAdd - Ticket added to pool: gren234-6 rolo
