--- RemasteredCLI/CLI/logs/ticketPool_save_tickets.log
@@ -1,29 +1,0 @@
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-3] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-3] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:20 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-4] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-5] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-6] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-7] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:21 [Thread-3] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:22 [Thread-2] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:22 [Thread-1] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:22 [Thread-0] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:22 [Thread-4] INFO  TicketPoolSave - Tickets saved to file.
-2024-11-23 17:36:22 [Thread-5] INFO  TicketPoolSave - Tickets saved to file.
