--- RemasteredCLI/CLI/logs/vendor_run.log
@@ -1,23 +1,0 @@
-2024-11-23 17:36:20 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-1 yolo
-2024-11-23 17:36:20 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-1 rolo
-2024-11-23 17:36:20 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-1 polo
-2024-11-23 17:36:20 [Thread-3] INFO  VendorRun - Ticket added by gren234: gren234-1 holo
-2024-11-23 17:36:20 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-2 polo
-2024-11-23 17:36:20 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-2 yolo
-2024-11-23 17:36:20 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-2 rolo
-2024-11-23 17:36:20 [Thread-3] INFO  VendorRun - Ticket added by gren234: gren234-2 holo
-2024-11-23 17:36:20 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-3 polo
-2024-11-23 17:36:20 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-3 rolo
-2024-11-23 17:36:20 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-3 yolo
-2024-11-23 17:36:21 [Thread-3] INFO  VendorRun - Ticket added by gren234: gren234-3 holo
-2024-11-23 17:36:21 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-4 polo
-2024-11-23 17:36:21 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-4 yolo
-2024-11-23 17:36:21 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-4 rolo
-2024-11-23 17:36:21 [Thread-3] INFO  VendorRun - Ticket added by gren234: gren234-4 holo
-2024-11-23 17:36:21 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-5 polo
-2024-11-23 17:36:21 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-5 yolo
-2024-11-23 17:36:21 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-5 rolo
-2024-11-23 17:36:21 [Thread-3] INFO  VendorRun - Ticket added by gren234: gren234-5 holo
-2024-11-23 17:36:22 [Thread-2] INFO  VendorRun - Ticket added by jren234: jren234-6 yolo
-2024-11-23 17:36:22 [Thread-1] INFO  VendorRun - Ticket added by kren234: kren234-6 polo
-2024-11-23 17:36:22 [Thread-0] INFO  VendorRun - Ticket added by gren234: gren234-6 rolo
