--- RemasteredCLI/CLI/src/Main.java
@@ -1,291 +1,0 @@
-import java.util.List;
-import java.util.Scanner;
-import java.util.Set;
-
-public class Main {
-    private static final Scanner input = new Scanner(System.in);
-    private static Configuration config = new Configuration().loadConfiguration("config.json");
-    private static TicketPool ticketPool = TicketPool.getInstance(config.getMaxTicketCapacity());
-
-    public static void main(String[] args) {
-        System.out.println("Welcome to Real-time Ticketing System by HKR\n");
-        while (true) {
-            ticketPool.setMaxCapacity(config.getMaxTicketCapacity());
-            System.out.println(ticketPool.getMaxCapacity());
-            printMainMenu();
-            int option = getUserChoice(1, 8);
-            if (option == 8) break;
-            handleMainOption(option);
-        }
-        System.out.println("Exiting System...");
-    }
-
-    private static void printMainMenu() {
-        System.out.println("Pick an option below:");
-        System.out.println("1. Configure Settings");
-        System.out.println("2. Login to System");
-        System.out.println("3. View Available Tickets");
-        System.out.println("4. View Real-time Tickets Being Sold and Added");
-        System.out.println("5. Stimulate ticket operations in the ticket pool");
-        System.out.println("6. Start System if Stopped");
-        System.out.println("7. Stop System (Stop All Ticket Operations)");
-        System.out.println("8. Exit System");
-        System.out.println();
-    }
-
-    private static int getUserChoice(int min, int max) {
-        while (true) {
-            try {
-                System.out.print("Enter your choice: ");
-                int choice = Integer.parseInt(input.nextLine().trim());
-                System.out.println();
-                if (choice >= min && choice <= max) return choice;
-                System.out.printf("Please enter a number between %d and %d.%n%n", min, max);
-            } catch (NumberFormatException e) {
-                System.out.println("Invalid input. Please enter a valid integer.\n");
-            }
-        }
-    }
-
-    private static void handleMainOption(int option) {
-        switch (option) {
-            case 1 -> handleConfigurationSettings();
-            case 2 -> handleLogin();
-            case 3 -> {
-                System.out.println("Viewing Available Tickets...\n");
-                ticketPool.countAvailableTicketsByEvent();
-                System.out.println();
-            }
-            case 4 -> System.out.println("Viewing Real-time Tickets Being Sold and Added...\n");
-            case 5 -> {
-                System.out.println("Stimulating ticket operations in the ticket pool...\n");
-                VendorTest stimulateVendor = new VendorTest(ticketPool);
-                stimulateVendor.startBotRelease();
-                try {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
-                StimulateCustomers stimulatecustomers = new StimulateCustomers(ticketPool);
-                stimulatecustomers.startBotPurcahse();
-            }
-            case 6 -> {
-                System.out.println("Starting System if Stopped...\n");
-                Vendor.resumeAllReleases();
-                Customer.resumeAllPurchases();
-            }
-            case 7 -> {
-                System.out.println("Stopping System (All Ticket Operations Halted)...\n");
-                Vendor.stopAllReleases();
-                Customer.stopAllPurchases();
-            }
-            default -> System.out.println("Invalid option. Please try again.\n");
-        }
-    }
-
-    private static void handleConfigurationSettings() {
-        System.out.println("Configuration Settings:");
-        System.out.println("1. Edit Admin credentials");
-        System.out.println("2. Edit Configuration Setting");
-        System.out.println("3. View Configuration Setting");
-        System.out.println("4. Exit");
-        int choice = getUserChoice(1, 4);
-        switch (choice) {
-            case 1 -> {System.out.println("Editing Admin Credentials...\n");
-                config.setAdminCredentials();
-                System.out.println("\n");
-            }
-            case 2 -> {System.out.println("Editing Configuration Setting...\n");
-                config.promptUserForConfigurationUpdate();
-                System.out.println("\n");
-            }
-            case 3 -> {System.out.println("Viewing Configuration Setting...\n");
-                config.displayConfiguration();
-                System.out.println("\n");
-            }
-            case 4 -> System.out.println("Exiting Configuration Settings...\n");
-        }
-    }
-
-    private static void handleLogin() {
-        System.out.println("Login:");
-        System.out.println("1. Login as a Vendor");
-        System.out.println("2. Login as a Customer");
-        System.out.println("3. Exit");
-        int choice = getUserChoice(1, 3);
-        switch (choice) {
-            case 1 -> handleVendorLogin();
-            case 2 -> handleCustomerLogin();
-            case 3 -> System.out.println("Exiting Login Menu...\n");
-        }
-    }
-
-    private static void handleVendorLogin() {
-        System.out.println("Vendor Login:");
-        System.out.println("1. Sign up Vendor");
-        System.out.println("2. Sign in Vendor");
-        System.out.println("3. Exit");
-        int choice = getUserChoice(1, 3);
-        switch (choice) {
-            case 1 -> {System.out.println("Signing up Vendor and Saving...\n");
-               boolean signup = Vendor.signUp();
-               if(signup){
-                   System.out.println("Sign up was successful!\n");
-               }else{
-                   System.out.println("Vendor Sign up was unsuccessful!\n");
-               }
-            }
-            case 2 -> {
-                Vendor vendor = Vendor.signIn();
-                if (vendor != null){
-                handleVendorActions(vendor);}
-            }
-            case 3 -> System.out.println("Exiting Vendor Login...\n");
-        }
-    }
-
-    private static void handleVendorActions(Vendor vendor) {
-        vendor.setTicketReleaseRate();
-        vendor.setTicketPool(ticketPool);
-        boolean exit = false;
-        while (!exit) {
-            System.out.println("Vendor Actions:");
-            System.out.println("1. Add Tickets (Create Threads)");
-            System.out.println("2. View Added Tickets of the Vendor");
-            System.out.println("3. View Other Tickets");
-            System.out.println("4. Stop Ticket Release for the Vendor");
-            System.out.println("5. Activating Ticket Release for the Vendor");
-            System.out.println("6. View Real-time Tickets Being Sold and Added");
-            System.out.println("7. Exit");
-            int choice = getUserChoice(1, 7);
-            exit = handleVendorChoice(choice, vendor);
-        }
-    }
-
-    private static boolean handleVendorChoice(int choice, Vendor vendor) {
-        switch (choice) {
-            case 1 -> {
-                System.out.println("Adding Tickets...\n");
-                if (vendor.checkReleaseStatus()) {
-                    List<Ticket> tickets = Ticket.createTicketsForVendor(vendor.getVendorId());
-                    vendor.setTicketBatch(tickets);
-                    Thread vendorThread = new Thread(vendor);
-                    vendorThread.start();
-                }
-            }
-            case 2 -> {
-                System.out.println("Viewing Vendor's Added Tickets...\n");
-                ticketPool.viewAvailableTicketCountsByVendor(vendor.getVendorId());
-                System.out.println();
-            }
-            case 3 -> {
-                System.out.println("Viewing Other Tickets...\n");
-                ticketPool.countAvailableTicketsByEvent();
-                System.out.println();
-            }
-            case 4 -> {
-                System.out.println("Stopping Ticket Release...\n");
-                vendor.stopReleasingTickets();
-            }
-            case 5 -> {
-                System.out.println("Activating Ticket Release...\n");
-                vendor.resumeReleasingTickets();
-            }
-            case 6 -> System.out.println("Viewing Real-time Tickets Being Sold and Added...\n");
-            case 7 -> {
-                System.out.println("Exiting Vendor Actions...\n");
-                return true;
-            }
-            default -> System.out.println("Invalid option. Please try again.\n");
-        }
-        return false;
-    }
-
-    private static void handleCustomerLogin() {
-        System.out.println("Customer Login:");
-        System.out.println("1. Sign up Customer");
-        System.out.println("2. Sign in Customer");
-        System.out.println("3. Exit");
-        int choice = getUserChoice(1, 3);
-        switch (choice) {
-            case 1 -> {
-                System.out.println("Signing up Customer and Saving...\n");
-                boolean signup = Customer.signUp();
-                if(signup){
-                    System.out.println("Sign up was successful!\n");
-                }else{
-                    System.out.println("Customer Sign up was unsuccessful!\n");
-                }
-            }
-            case 2 -> {
-                Customer customer = Customer.signIn();
-                if(customer != null){
-                handleCustomerActions(customer);}
-            }
-            case 3 -> System.out.println("Exiting Customer Login...\n");
-        }
-    }
-
-    private static void handleCustomerActions(Customer customer) {
-        customer.setCustomerRetrievalRate();
-        customer.setTicketPool(ticketPool);
-        boolean exit = false;
-        while (!exit) {
-            System.out.println("Customer Options:");
-            System.out.println("1. Purchase Tickets");
-            System.out.println("2. View Tickets Bought");
-            System.out.println("3. View Other Tickets");
-            System.out.println("4. Stop purchase of Tickets for the customer");
-            System.out.println("5. Activate purchase of Tickets for the customer");
-            System.out.println("6. View Real-time Tickets Being Sold and Added");
-            System.out.println("7. Exit");
-            int choice = getUserChoice(1, 7);
-            exit = handleCustomerChoice(choice, customer);
-        }
-    }
-
-    private static boolean handleCustomerChoice(int choice, Customer customer) {
-        switch (choice) {
-            case 1 -> {
-                System.out.println("Purchasing Tickets...\n");
-                if (customer.checkPurchaseStatus()) {
-                    Set<String> eventDetails = ticketPool.getEventDetails();  // Fetch valid event names
-                    PurchaseRequest request = customer.requestTickets(eventDetails);
-                    if (request != null) {
-                        customer.setPurchaseRequest(request);
-                        Thread customerThread = new Thread(customer);
-                        customerThread.start();
-                    } else {
-                        System.out.println("No purchase request available to process.\n");
-                    }
-                }
-            }
-            case 2 -> {
-                System.out.println("Viewing Tickets Bought...\n");
-                ticketPool.countBookedTicketsByCustomerId(customer.getCustomerId());
-                System.out.println();
-            }
-            case 3 -> {
-                System.out.println("Viewing Other Tickets...\n");
-                ticketPool.countAvailableTicketsByEvent();
-                System.out.println();
-            }
-            case 4 -> {
-                System.out.println("Stopping purchase of Tickets for the customer...\n");
-                customer.stopPurchasingTickets();
-            }
-            case 5 -> {
-                System.out.println("Activating purchase of Tickets for the customer...\n");
-                customer.resumePurchasingTickets();
-            }
-            case 6 -> System.out.println("Viewing Real-time Tickets Being Sold and Added...\n");
-            case 7 -> {
-                System.out.println("Exiting Customer Actions...\n");
-                return true;
-            }
-            default -> System.out.println("Invalid option. Please try again.\n");
-        }
-        return false;
-    }
-}
-
