--- RemasteredCLI/README.md
@@ -1,2 +1,0 @@
-# CW_OOP_w2082289
-This is the repo that will be used for version control of the OOP coursework
